% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bci_2010.R
\docType{data}
\name{bci_2010}
\alias{bci_2010}
\title{bci_2010}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 299 rows and 1250 columns.
}
\source{
\url{https://repository.si.edu/handle/10088/20925}
}
\usage{
bci_2010
}
\description{
The 2010 census from the Barro Colorado Island (BCI) dataset,
containing a count og known species of living trees in a 20x20m quadrats.
}
\keyword{datasets}
