% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bci_quadrats.R
\docType{data}
\name{bci_quadrats}
\alias{bci_quadrats}
\title{bci_quadrats}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1250 rows and 6 columns.
}
\source{
\url{https://repository.si.edu/handle/10088/20925}
}
\usage{
bci_quadrats
}
\description{
Contains metadata about the quadrats, such as where they are
}
\keyword{datasets}
