% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bci_taxa.R
\docType{data}
\name{bci_taxa}
\alias{bci_taxa}
\title{bci_taxa}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 299 rows and 3 columns.
}
\source{
\url{https://repository.si.edu/handle/10088/20925}
}
\usage{
bci_taxa
}
\description{
Tiblle containing the species name (GenusSpecies),
genus(Genus) and family (Family) of each tree specie
on Barro Colorado Island (BCI)
}
\keyword{datasets}
