% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rory646BCI-package.R
\docType{package}
\name{Rory646BCI-package}
\alias{Rory646BCI-package}
\alias{Rory646BCI}
\title{Rory646BCI}
\description{
\code{Rory646BCI} is a dummy package developed as part of the Reproducible
Programming in R course..
}
\details{
This text will appear in the Details section of the documentation.
}
\author{
Rory <2188096T@student.gla.ac.uk> (maintainer) \cr
}
