% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_by_subcommunities.R
\name{sample_by_subcommunities}
\alias{sample_by_subcommunities}
\title{Subsample a dataset by subcommunities}
\usage{
sample_by_subcommunities(dataset, count)
}
\arguments{
\item{dataset}{A matrix, data frame or tibble containing abundance or incidence data}

\item{count}{The number of subcommunities to retain}
}
\value{
The subsampled dataset in the format it was passed in
}
\description{
Sample a dataset as if we only counted in some subcommunities.
}
\examples{
library(magrittr)
sample_by_species(bci_2010, count = 20) \%>\%
  sample_by_subcommunities(count = 6)

}
