% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rory646BCI-package.R
\docType{package}
\name{Rory646BCI-package}
\alias{Rory646BCI-package}
\alias{Rory646BCI}
\title{Barro Colorado Island data package}
\description{
Package for Series 4 of the RPiR course, created to hold BCI
data and functions created to analyse diversity.
}
